sidecar/requirements.txt
python-telegram-bot==21.10
apscheduler==3.11.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
httpx==0.28.1
fastapi==0.115.12
uvicorn==0.34.2
websocket-client==1.8.0
